Update for Swift 2.1
Ditch the `newAssocKey()` function in favor of initializers on
`AssocKey` itself, using extensions with `where` clauses to constrain
them appropriately.

Convert `AssocKey` into a `struct` and use an underlying private class
instance as the actual key. This way leaked keys will just leak memory
instead of crashing.

Fix a safety hole where a generic wrapper could call `get()` or `set()`
with a value that conforms to `AnyObject` at runtime but not
compile-time. We were relying on the overloaded function to provide type
safety, but in a generic wrapper like that the compiler will always call
the non-overloaded version like that. Resort to runtime type-checking
that will hopefully be optimized out. This should also mean that strings
and other bridgeable types will still be stored as values instead of
bridging to obj-c first.