Invert the API, change AssocKey to AssocMap
Associated objects are semantically just a map from `AnyObject` to
`ValueType`, where the keys are held weakly (but don't require weak
support), and the values are discarded immediately whenever the key
deallocs. We can actually provide an API that looks like this, which has
the benefits of letting us use the subscript operator.

Get rid of the compile-time overload on the getter/setter for
`ValueType: AnyObject`. It occurs to me that this overload would be used
for bridgeable types even though the generic version (which tests
`ValueType.self is AnyObject`) intentionally doesn't deal with obj-c
bridging, which means the overload is actually unsafe in certain
contexts. The optimizer really should be able to get rid of the
type-check anyway.